pyproject.toml
[tool.poetry]
name = "shipment_management"
version = "0.1.0"
description = "Generic integration with multiple shipment providers"
authors = ["Abdelrahman Elsharawy"]

[tool.poetry.dependencies]
python = "3.8.12"
Django = "4.1.1"
djangorestframework = "^3.13.1"
django-extensions = "^3.2.0"
celery = "^5.2.7"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
